Refactor progress bar/display into callbacks
Split up the UI code and the download- and parsing code into separate functions instead of single functions that deal with both.
The download- and parse functions now take callbacks that they use to display progress. The old behavior is put in default callbacks
so there is no change for users.

Rename `count` to `length`