More friendly support for highdim constraints
getHighdimData now has support for constraints with a customizable
syntax, as the rest-api encoding sometimes uses arcane named parameters.

Also switch RJSONIO to rjson. According to the docs there shouldn't be
much of a speed difference, but rjson gives a tighter encoding without
extraneous whitespace, which is nicer when json is encoded into url
parameters.