Make constraint names more idiomatic R
Add a general renaming mechanism for the constraint names, so now we
use e.g. ontology.term instead of ontology_term.
Also homogenize the datatype handling. Numbers are checked with
is.numeric(), and characters are coerced with as.character(). It is
up to the user not to pass fractional numbers where those are not
allowed, but integral valued doubles are handled fine.